Upgrading projects to ASP.NET MVC4 and factored out database
Upgrading projects to ASP.NET MVC4 and factored out database

+++ src/MvcSiteMapProvider/MvcMusicStore/ViewModels/ShoppingCartRemoveViewModel.cs
@@ -8,7 +8,7 @@ namespace MvcMusicStore.ViewModels
     public class ShoppingCartRemoveViewModel
     {
         public string Message { get; set; }
-        public decimal CartTotal { get; set; }
+        public double CartTotal { get; set; }
         public int CartCount { get; set; }  
         public int DeleteId { get; set; }
     }
