+++ src/MvcSiteMapProvider/MvcMusicStore/ViewModels/ShoppingCartViewModel.cs
@@ -6,6 +6,6 @@ namespace MvcMusicStore.ViewModels
     public class ShoppingCartViewModel
     {
         public List<Cart> CartItems { get; set; }
-        public decimal CartTotal { get; set; }
+        public double CartTotal { get; set; }
     }
 }
